requirements.txt
streamlit
pytesseract
pdf2image
Pillow
opencv-python
